feat(render): add ParticleData and ParticleQueueManager for particle management
